Fixing problems, creating writers package in metron common, and moving messagewriter classes there
