+++ metron-platform/metron-common/src/main/java/org/apache/metron/common/writer/BulkMessageWriter.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.metron.common.interfaces;
+package org.apache.metron.common.writer;
 
 import backtype.storm.tuple.Tuple;
 import org.apache.metron.common.configuration.writer.WriterConfiguration;
