+++ metron-platform/metron-common/src/main/java/org/apache/metron/common/writer/test.json
@@ -1,0 +1,28 @@
+{
+  "index": "bro",
+  "batchSize": 1,
+  "enrichment" : {
+    "fieldMap":
+    {
+      "geo": ["ip_dst_addr", "ip_src_addr"],
+      "host": ["host"]
+    }
+  },
+  "threatIntel" : {
+    "fieldMap":
+    {
+      "hbaseThreatIntel": ["ip_src_addr", "ip_dst_addr"]
+    },
+    "fieldToTypeMap":
+    {
+      "ip_src_addr" : ["malicious_ip"],
+      "ip_dst_addr" : ["malicious_ip"]
+    },
+    "triageConfig" : {
+      "riskLevelRules" : {
+        "ip_src_addr == '31.24.30.31'" : "Test"
+      },
+      "aggregator" : "MAX"
+    }
+  }
+}
