+++ metron-platform/metron-writer/src/main/java/org/apache/metron/writer/WriterToBulkWriter.java
@@ -22,9 +22,9 @@ import backtype.storm.tuple.Tuple;
 import com.google.common.collect.Iterables;
 import org.apache.metron.common.configuration.writer.SingleBatchConfigurationFacade;
 import org.apache.metron.common.configuration.writer.WriterConfiguration;
-import org.apache.metron.common.interfaces.BulkMessageWriter;
-import org.apache.metron.common.interfaces.MessageWriter;
-import org.apache.metron.common.interfaces.BulkWriterResponse;
+import org.apache.metron.common.writer.BulkMessageWriter;
+import org.apache.metron.common.writer.MessageWriter;
+import org.apache.metron.common.writer.BulkWriterResponse;
 
 import java.io.Serializable;
 import java.util.List;
@@ -46,12 +46,19 @@ public class WriterToBulkWriter<MESSAGE_T> implements BulkMessageWriter<MESSAGE_
 
   @Override
   public BulkWriterResponse write(String sensorType, WriterConfiguration configurations, Iterable<Tuple> tuples, List<MESSAGE_T> messages) throws Exception {
-    if(messages.size() > 1) {
-      throw new IllegalStateException("WriterToBulkWriter expects a batch of exactly 1");
-    }
-    messageWriter.write(sensorType, configurations, Iterables.getFirst(tuples, null), Iterables.getFirst(messages, null));
-
     BulkWriterResponse response = new BulkWriterResponse();
+    if(messages.size() > 1) {
+        response.addAllErrors(new IllegalStateException("WriterToBulkWriter expects a batch of exactly 1"), tuples);
+        return response;
+    }
+
+    try {
+      messageWriter.write(sensorType, configurations, Iterables.getFirst(tuples, null), Iterables.getFirst(messages, null));
+    } catch(Exception e) {
+      response.addAllErrors(e, tuples);
+      return response;
+    }
+
     response.addAllSuccesses(tuples);
     return response;
   }
